requirements.txt
Werkzeug
flask
peewee
python-dotenv
pymysql
Pillow
discord.py
requests
awesome-slugify
flask-uploads